Added the ability to add price overrides
