+++ frontend/src/Components/Pages/AdminPanel/Panels/ItemUpdatePanel/ItemUpdatePanel.scss
@@ -1,0 +1,96 @@
+.disabled-items {
+  position: relative;
+  padding: 10px;
+  font-family: 'Poppins';
+
+  .StoreContent {
+    border-radius: 7.55px;
+    height: 79vh;
+    overflow-y: scroll;
+    background-color: #dadada;
+    padding: 10px;
+    max-width: 90%;
+
+    .item {
+      display: flex;
+      flex-wrap: wrap;
+      align-content: flex-start;
+      
+      .price-disable {
+        display: flex;
+        margin-top: 15px;
+        width: 100%;
+        border-radius: 7.55px;
+
+        .price-input {
+          margin-right: 5px;          
+        }
+
+        .price-update {
+          margin-left: 5px;
+          margin-right: 5px;
+          border-radius: 7.55px;
+        }
+
+        .price-delete {
+          margin-left: 5px;
+        }
+      }
+
+      .price-disable-add {
+        margin-top: 10px;
+        width: 100%;
+        border-radius: 7.55px;
+      }
+    }
+  }
+
+  .top-bar {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+    width: 90%;
+    margin: auto;
+
+    .search {
+      margin: 5px;
+    }
+
+    .mass-price-override {
+      margin: 5px;
+      display: flex;
+      flex-wrap: wrap;
+      margin-right: 10px;
+      margin-left: 10px;
+      border-radius: 7.55px;
+
+      .mass-price-button {
+        margin: 5px;
+      }
+
+    }
+
+    .toggle-button {
+      margin: 5px;
+      border-radius: 7.55px;
+      // font-size: 30px;
+    }
+  }
+
+  .selected {
+    border: 10px solid rgb(255, 0, 170);
+  }
+
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: column;
+  width: 100%;
+  text-align: left;
+
+  padding: 5px;
+
+  h1.title {
+    font-size: 50px;
+    color: white;
+  }
+}
