+++ frontend/src/Components/Pages/AdminPanel/Panels/index.js
@@ -1,9 +1,9 @@
 import CommandsPanel from './CommandsPanel/CommandsPanel';
-import ItemDisablePanel from './ItemDisablePanel/ItemDisablePanel';
+import ItemUpdatePanel from './ItemUpdatePanel/ItemUpdatePanel';
 import PlayerManagePanel from './PlayerManagePanel/PlayerManagePanel';
 
 export {
   CommandsPanel,
-  ItemDisablePanel,
+  ItemUpdatePanel,
   PlayerManagePanel,
 };
